requirements.txt
data==0.4
matplotlib==3.4.3
numpy>=1.18.1
Pillow==9.1.1
scipy==1.7.1
torch==1.5.0
torchvision==0.6.0
